pyproject.toml
[project]
name = "prisam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx>=0.2.4",
    "google-generativeai>=0.8.5",
    "groq>=0.33.0",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.2",
    "langchain-groq>=1.0.0",
    "langgraph>=1.0.2",
    "nltk>=3.9.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "pymupdf>=1.26.5",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
